Clean up some signed vs. unsigned warnings.
Octet arrays are octets of guint8s, not gchars/chars.

Make some strings arrays of chars/gchars, not guint8s; this needs more
thought (throughout Wireshark).

Offsets into tvbuffs are signed, not unsigned.  (This is to support
negative offsets, which are offsets from the end of the tvbuff.  We
might want to remove that and go with unsigned offsets, and have the
few, if any, places where that feature is used explicitly calculate the
offset from the end based on the tvbuff's length; most if not all of our
handling of trailers/end-of-packet FCSes/etc. does so, and makes sure it
handles the case where the end-of-packet information isn't present, to
better report errors and dissect the stuff before it.)

Change-Id: Ia46ed3fc7c2d8ac97cd14824d521cbc461fb7f45
Reviewed-on: https://code.wireshark.org/review/33239
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>