"Zigbee" encoding is not a character encoding or an integer encoding.
It's a rule for interpretation of the length field of counted-string and
counted-octet-string fields.  This means it's 1) not a general rule for
interpreting integers and 2) not a character encoding, as it also
applies to octet strings and, even for character strings, it's
*orthogonal* to the character encoding.

Therefore, it should *not* be one of the character encoding values; it
should be a bit flag.

Make it so.  This means that

1) a character encoding can be specified for Zigbee Cluster Library
strings (they appear to have multiple character encodings possible);

2) the test of it that tested it as if it were a flag will no longer get
confused by character encodings that set one or more of the bits in the
old encoding value;

3) you don't have to special-case the encoding value passed to
get_uint_value().

Put in a comment emphasizing that values that aren't character encodings
should *not* be placed in the set of character encodings.

Change-Id: I8f50aaee8ca60b0781044287e9b38111de38c81f
Reviewed-on: https://code.wireshark.org/review/33341
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>