*Little-endian* UTF-16.
ENC_UTF_16 does *not* go with ENC_NA; ENC_NA is for cases where the byte
order is "not applicable", such as a 1-byte number or a character
encoding where every character is encoded in 1 byte, but UTF-16 isn't
one of those cases, as a character is encoded in either 1 or 2 2-byte
values.  This being a Windows thing, the byte order is little-endian.

Change-Id: Iab0db3fa2c5d2c25be209e4ed0ebd57827edbcd8
Reviewed-on: https://code.wireshark.org/review/33347
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 2114dba1effebba50cdb611b2650b0a4a37761e4)
Reviewed-on: https://code.wireshark.org/review/33349