Clean up some ASCII vs. EBCDIC string handling.
In at least one capture, structure IDs are in ASCII even though the code
page in the header is an EBCDIC code page.  Determine the structure ID's
character encoding based on whether it's the ASCII or EBCDIC version of
the ID value, not on the global character encoding.

We were using the *integer* encoding, not the *string* encoding, for the
"qprotect" field, which is a string; fix that.

Use STR_UNICODE for strings, as they're not guaranteed to consist of
characters that can be mapped to ASCII characters (even the common
subset of EBCDIC, not counting code page-dependent code points, has
non-ASCII printable characters in it).

Change-Id: I971dd7ae55617c27ebe88f31089b2495374593bf
Reviewed-on: https://code.wireshark.org/review/33399
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(backported from commit b9c69d6ef8b2c759bb1b4be05240bba42038a051)
Reviewed-on: https://code.wireshark.org/review/33402