Reassemble PLP chunks.
Display the data in a PLP chunk as just raw bytes, and reassembled all
the chunks and extract the value from the reassembled data.

Showing the chunks as strings doesn't necessarily work - if the string
is an NVarChar string, a chunk might well have an odd number of octets,
meaning you're splitting one of the 16-bit UTF-16LE items in half.
Reassembling handles that, as well as just, in general, showing the
actual value rather than pieces of the value.

If the column is a string, append its value to the top-level item, just
as we do for non-PLP strings.

Also, if a length value was specified, report an error if it doesn't
match the total length of the reassembled chunks.

Change-Id: Iab80da052eb363ee08cd518afbe2556a5ab740b9
Reviewed-on: https://code.wireshark.org/review/33466
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>