Don't assume an NT Create AndX request has all the word parameters.
It *should*, but a malicious or otherwise malformed packet might not
have them.  One of them is the file name length; if it's missing, we
can't dissect the file name, as we don't know how long it is.

Change-Id: Ie259e2d8ec65f5d53d466382d89889902495d2c8
Reviewed-on: https://code.wireshark.org/review/33467
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit f0c13af7d3862f11b47cfcb12ccc80341122395d)
Reviewed-on: https://code.wireshark.org/review/33471