Handle some weirdness with the primary domain field in NegProt replies.
Sometimes there appears to be an extra byte before that field; try to
catch some of those cases.

Expand comments discussing various weirdness with that field, including
a note that clients might not pay any attention to it, so maybe we just
have buggy servers talking to clients that don't care about those
particular bugs.

Change-Id: I4d35d2e2c475d4da37debedfed31b891e6f3cfa8
Reviewed-on: https://code.wireshark.org/review/33481
Reviewed-by: Guy Harris <guy@alum.mit.edu>