batadv: Add support for ELP packets
The B.A.T.M.A.N. V protocol replaced the OGM announcement with two new
types:

* Echo Location Protocol packet
* OriGinator Message 2 packets

The first packet is used locally to identify neighbors and their base
parameters (orignator + elp interval). The second one is used to announce
each mesh node globally.

The first step to understand the B.A.T.M.A.N. V mesh locally in wireshark
is to dissect the ELP packets.

Change-Id: I13f5d60637a2774282ca75853884238e8e7dd33c
Signed-off-by: Sven Eckelmann <sven@narfation.org>
Reviewed-on: https://code.wireshark.org/review/33543
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Reviewed-by: Anders Broman <a.broman58@gmail.com>