Qt: traffic table: remove file_closed_
Remove the file_closed_ member from class TrafficTableDialog. We already
have a file_closed_ variable in the base class WiresharkDialog, which is
updated correctly when the file is closed.

The shadow file_closed_ variable in TrafficTableDialog is always false.
Classes that are derived from TrafficTableDialog will not see that the
capture file was closed.

The following scenario crashes Wireshark because of this bug

* open a capture file with TCP traffic
* Statistics / Conversations
* select a TCP conversation
* close the capture file
* press the Follow Stream button in the Conversations dialog
  -> Wireshark crashes

Change-Id: I8f3d55f231eae5ecc682ff90650f0005fea4b333
Reviewed-on: https://code.wireshark.org/review/33578
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>