Get rid of the nt_trans_data structure.
It was used for two purposes:

	1) passing the subcommand to dissectors for the request setup,
	   parameters, and data;

	2) getting the security descriptor and extended attribute length
	   from the request parameters and passing them to the dissector
	   for the request data.

It was completely unused for responses.

Pass the subcommand as an argument to all routines that need it,
pass pointers to local variables to the request parameters dissector,
and pass those variables to the request data dissector.

This makes it a bit clearer what information is needed where, and
completely eliminates the structure for response dissection.

Change-Id: I431daa79da41dc4166490a564efee71cdb83edc1
Reviewed-on: https://code.wireshark.org/review/33614
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>