f5ethtrailer - Dissect BIG-IP v15 low noise trailer
Bug: 15876

Add code to understand the version 3 "low noise" trailer.

Dissect fields added in the version 3 trailer.

Re-organize the "ingress" field from trailer version <= 2 to a flags field for
version >= 3.  Mark the f5ethtrailer.ingress field as a gerenated field in
version 3 and higher so that users can still use it for filtering, but move the
"real" ingress to be part of the flags.

Change-Id: I946e91e6b482ee5bd4e6d5a9c2523613a607c609
Reviewed-on: https://code.wireshark.org/review/33743
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>