Libpcap may now say "The interface disappeared" if it did.
A recent change to libpcap means that the error message if an interface
disappears (e.g., removing a hot-pluggable device, or shutting down a
PPP connection that was dynamically set up) is "The interface
disappeared" rather than "The interface went down" - on FreeBSD,
DragonFly BSD, OpenBSD, and Darwin-based OSes, capturing continues with
no error if the interface is configured down, but either ENXIO or EIO
(depending on the OS) is delivered if the interface disappears.

Treat that error as another one to show the user without the "report
this to the Wireshark developers" note.

Change-Id: I477d87957ce30a52385f07f4b47a7824e3fca2c7
Reviewed-on: https://code.wireshark.org/review/33790
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 8a4ce74ac9a3be8c752c1b405349d6083f76e657)
Reviewed-on: https://code.wireshark.org/review/33792