Qt: fix QIcon crash on exit on Ubuntu 16.04 with Qt 5.5.1
The icon used by DisplayFilterCombo is not cleaned up when MainWindow is
destroyed, so Qt cleans it up on program exit. Due to another Qt bug
(https://bugreports.qt.io/browse/QTBUG-50829, fixed in Qt 5.8.0), the
QIconEngine plugin that backs the QIcon was unloaded before the cleanup
was complete. As a result, calling the QIcon destructor would result in
calling the unmapped QIconEngine destructor code, leading to a
segmentation fault.

master-3.0 and later are not affected since a similar fix was added via
v2.9.0rc0-2728-gabfe9572bd.

Bug: 15241
Change-Id: Ie6bd8e645a2c754d6b57290e7084eaa324696253
Fixes: v2.6.2rc0-170-g3a0da5fbb2 ("Qt: Create the main display filter combo sooner.")
Reviewed-on: https://code.wireshark.org/review/33828
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>