Clean up handling of the command parameter.
It's an unsigned int, as per the protocol spec.  Treat it as such.  (Add
a URL for the protocol spec while we're at it.)

Convert it from a string of ASCII hex digits to a number using
ws_hexstrtou32(); that routine does the appropriate checks.

Clean up blank lines while we're at it - add some and remove some.

Change-Id: I2c0832c4a57c121e2e64de49a8678b747025f35e
Reviewed-on: https://code.wireshark.org/review/33972
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>