Don't just grab raw string data with tvb_memcpy().
Use proto_tree_add_item_ret_display_string() routines to add strings if
we want to display the string's value in a column.  That way, all
strings are fetched using an encoding value, to properly map to UTF-8,
and are formatted for display.

Change-Id: I4acd9ed7cfad3342be84a4773187dd531949f47b
Reviewed-on: https://code.wireshark.org/review/33974
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>