Use proto_tree_add_item_ret_display_string() to add string items.
It does all the work done by the calls it replaces, *and* it takes an
encoding argument, so the string data is converted, if necessary, to
UTF-8.  (Not necessary in this case, at least as I read the protocol
spec, but it's the right thing to do in any case.)

Plus, as we're using proto_tree_add_item_ret_display_string(), the
string we're displaying will be in a displayable format.

Change-Id: I7f5e46dcdbb8f3d3749ef539a31f6208371b72cd
Reviewed-on: https://code.wireshark.org/review/34022
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>