Clean up the handling of the "friendly name".
Fetch it by adding it with proto_tree_add_item_ret_display_string(), and
then use the resulting displayable string to append to columns and
protocol tree items.  Given that the string in question is ISO 8859-1,
according to the KNXnet/IP spec, and that it must therefore be converted
to UTF-8, that's the right thing to do.

Use wmem string buffers to hold the strings to append - using a
fixed-length buffer isn't a good idea when you are dealing with UTF-8
strings, as you might cut a UTF-8 sequence short in the middle.

Don't consruct strings that we never use.

While we're at it, give a URL to find KNX specifications.

Change-Id: Ibec4f6c83a62e141bd8ce0e5dfd7dd45ff627fe4
Reviewed-on: https://code.wireshark.org/review/34024
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>