genl: Always call subdissector
Commit 61c5e8e76d21 ("genl: make subdissectors responsible for header")
changed the generic netlink dissector to only call a sub-dissector if
there is a payload after the generic netlink header.

However, there are commands in certain generic netlink families that do
not have any payload. For example, 'NET_DM_CMD_START' in the 'NET_DM'
family. This means that the command will not be dissected by the
subdissector, as it will never be invoked.

Change the generic netlink dissector to always call a subdissector, if
it is present. Prevent the subdissectors from trying to dissect past the
end of the packet by adding checks in the two existing subdissectors,
for the 'nlctrl' and 'nl80211' families.

Change-Id: I4d2f48531dee92b11dc45000081a8d2d3dd875c6
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Reviewed-on: https://code.wireshark.org/review/34350
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>