f5ethtrailer: type-version subdissector table
Add a subdissector table to the noise provider.
This allows for extendability via a plugin to support
new versions or message types of trailer noise.

The original f5ethtrailer was distrubuted as a plugin.
Since including it as part of libwireshark it has been
impossible to add support for changes in the data this
dissector decodes once Wireshark is released without patching
and compiling on your own (or possibly building a git dev branch).

This provides a mechanisim to add additional support to the
dissector via plugin without needing to rebuild Wireshark.

Change-Id: I1912ab497e0cfd34dcd92a893b819333665e23cd
Reviewed-on: https://code.wireshark.org/review/34063
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>