Docbook: Document using middle-click to mark packets.
Add an item about marking packets using the middle mouse button to the
release notes.

Update the "Marking Packets" section of the User's Guide accordingly.

Use "menu:...[]" to mark up menu items in a bunch of places. It looks
like we need to a add a "guimenu" class to ws.css.

Change-Id: Ide99112f7643e509d8af8a4aa6ddb4287f3585cf
Reviewed-on: https://code.wireshark.org/review/34182
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>